+++ src/libs/gui_tk/gui_tk.h
@@ -2082,7 +2082,7 @@ protected:
             selected++;
         }
 
-        if(selected > items.size() - 1)
+        if(selected > (int)items.size() - 1)
             selected = items.size() - 1;
 
         if(selected >= 0 && items[(unsigned int)selected].size() == 0) selected = -1;
